refactor(180): update architecture for StudentSkillsWidgetContainer
