+++ src/features/student/components/containers/StudentSkillsWidgetContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as StudentSkillsWidgetContainer } from './StudentSkillsWidgetContainer.vue'
