+++ src/features/student/components/containers/index.ts
@@ -1,1 +1,1 @@
-export { default as StudentSkillsWidgetContainer } from './StudentSkillsWidgetContainer.vue'
+export * from './StudentSkillsWidgetContainer'
